+++ mineinabyss-components/src/main/kotlin/com/mineinabyss/components/mobtags/Insect.kt
